regress: adapt all C code to latest change in uc_hook_add() API

+++ tests/regress/block_test.c
@@ -65,7 +65,7 @@ int main() {
    
    uc_hook h1, h2;
 
-   err = uc_hook_add(uc, &h1, UC_HOOK_BLOCK, cb_hookblock, NULL, (uint64_t)1, (uint64_t)0);
+   err = uc_hook_add(uc, &h1, UC_HOOK_BLOCK, cb_hookblock, NULL, 1, 0);
    if (err != UC_ERR_OK) {
       fprintf(stderr, "not ok %d - %s\n", count++, uc_strerror(err));
       exit(0);
